# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API on Golang for film library service.",
        "title": "Film library service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/film/find": {
            "get": {
                "description": "Find film by title or actor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search field",
                        "name": "search_field",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Searched value",
                        "name": "value",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/find_film.FilmData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    }
                }
            }
        },
...
        "/user/login": {
            "post": {
                "description": "Login in service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login_user.UserDataIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/login_user.UserDataOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Register in service",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/register_user.UserDataIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "common.ErrorOut": {
            "type": "object",
            "properties": {
                "error_message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "find_film.FilmData": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Description"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "rating": {
                    "type": "integer",
                    "example": 10
                },
                "release_date": {
                    "type": "string",
                    "example": "1998-02-20"
                },
                "title": {
                    "type": "string",
                    "example": "Titanic"
                }
            }
        },
...
        "login_user.UserDataIn": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "username": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "login_user.UserDataOut": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "token"
                }
            }
        },
        "register_user.UserDataIn": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}